+++ lib/vagrant/environment.rb
@@ -302,13 +302,15 @@ module Vagrant
         @config_loader.load([:default, :home, :root, vm_config_key])
 
       box = nil
-      begin
-        box = boxes.find(config.vm.box, provider)
-      rescue Errors::BoxUpgradeRequired
-        # Upgrade the box if we must
-        @logger.info("Upgrading box during config load: #{config.vm.box}")
-        boxes.upgrade(config.vm.box)
-        retry
+      if config.vm.box
+        begin
+          box = boxes.find(config.vm.box, provider)
+        rescue Errors::BoxUpgradeRequired
+          # Upgrade the box if we must
+          @logger.info("Upgrading box during config load: #{config.vm.box}")
+          boxes.upgrade(config.vm.box)
+          retry
+        end
       end
 
       # If a box was found, then we attempt to load the Vagrantfile for